+++ exit.c
@@ -5,8 +5,8 @@
 #include "tasks.h"
 #include "key_input.h"
 
-static int exit_task;
-static int exit_now;
+static short exit_task;
+static short exit_now;
 static const char exit_help[] = "esc or Enter - end program";
 
 void exit_enable(void)
